Refactor dialogue parsing to use FastJSON

+++ src/main/java/com/raikuman/troubleclub/command/TestDialogue.java
@@ -2,22 +2,36 @@ package com.raikuman.troubleclub.command;
 
 import com.raikuman.botutilities.invocation.context.CommandContext;
 import com.raikuman.botutilities.invocation.type.Command;
-import com.raikuman.troubleclub.dialogue.DialogueManager;
+import com.raikuman.troubleclub.conversation.Conversation;
+import com.raikuman.troubleclub.conversation.ConversationManager;
+import com.raikuman.troubleclub.dialogue.Dialogue;
+import com.raikuman.troubleclub.dialogue.DialoguePlayer;
+import com.raikuman.troubleclub.dialogue.DialogueParser;
 
+import java.io.File;
 import java.util.List;
 
 public class TestDialogue extends Command {
 
-    private final DialogueManager dialogueManager;
+    private final ConversationManager conversationManager;
 
-    public TestDialogue(DialogueManager dialogueManager) {
-        this.dialogueManager = dialogueManager;
+    public TestDialogue(ConversationManager conversationManager) {
+        this.conversationManager = conversationManager;
     }
 
     @Override
     public void handle(CommandContext ctx) {
+        File file = new File("resources/conversations/sande.json");
+        Conversation conversation = DialogueParser.getConversation(file.toPath());
+        if (conversation == null) return;
+
+        Dialogue dialogue = conversation.getDialogue();
+
+        if (dialogue != null) {
+            DialoguePlayer.setup(conversationManager.getClubMap(), dialogue).play(null);
+        }
+
         ctx.event().getMessage().delete().queue();
-        dialogueManager.testDialogue();
     }
 
     @Override
